services/admin-service/requirements.txt
fastapi==0.131.0
bcrypt==4.1.2
slowapi==0.1.9
uvicorn[standard]==0.27.0
pydantic[email]==2.6.0
pydantic-settings==2.1.0
sqlalchemy==2.0.46
asyncpg==0.31.0
pyjwt==2.8.0
pino==1.7.0
pino-http==8.3.0
opentelemetry-api==1.22.0
opentelemetry-sdk==1.22.0
opentelemetry-exporter-otlp==1.22.0
opentelemetry-exporter-otlp-proto-http==1.22.0
opentelemetry-instrumentation-fastapi==0.43b0
opentelemetry-instrumentation-sqlalchemy==0.43b0
opentelemetry-instrumentation-redis==0.43b0
opentelemetry-instrumentation-requests==0.43b0
redis==5.0.1
httpx==0.26.0
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
coverage==7.4.0
prometheus-client==0.19.0
psutil==5.9.8
prometheus-fastapi-instrumentator==6.1.0
python-multipart==0.0.9
